+++ src/components/profile/CCEditProfileCard.tsx
@@ -3,7 +3,7 @@
 import { useModal } from "@/hooks/useModal";
 import DefaultButton from "@/components/ui/defaultButton";
 import EditSVG from "/public/icons/edit.svg";
-import CCProfileModal from "@/components/profile/CCProfileModal";
+import CCProfileEditModal from "@/components/profile/CCProfileEditModal";
 
 export default function CCEditProfileCard() {
   const { setIsOpenModal, isOpenModal, modalOutsideRef } = useModal();
@@ -17,7 +17,7 @@ export default function CCEditProfileCard() {
         프로필 수정
       </DefaultButton>
       {isOpenModal && (
-        <CCProfileModal
+        <CCProfileEditModal
           closeModal={() => setIsOpenModal(false)}
           modalRef={modalOutsideRef}
         />
